change structure and add const color
