+++ lib/screens/main_screen/widgets/clipper.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class Clipper extends CustomClipper<Path> {
+  @override
+  Path getClip(Size size) {
+    return Path()
+      ..moveTo(size.width, 150)
+      ..lineTo(0, size.height)
+      ..lineTo(size.width, size.height)
+      ..close();
+  }
+
+  @override
+  bool shouldReclip(covariant CustomClipper<Path> oldClipper) {
+    return false;
+  }
+}
