+++ lib/utils/colors_constants.dart
@@ -1,8 +1,10 @@
-import 'package:flutter/animation.dart';
+import 'package:flutter/material.dart';
 
 class ConstantsColors {
   ConstantsColors._();
 
+  static const defaultTextColor = Colors.white;
+  static const backgroundColor = Color.fromARGB(255, 53, 53, 53);
   static const deadpoolColor = Color.fromARGB(255, 129, 8, 8);
   static const ironManColor = Color.fromARGB(255, 250, 75, 75);
   static const captainAmericaColor = Color.fromARGB(255, 54, 82, 236);
